Extract profile edit coordinator
